Rework patterns to sagas example
