+++ patterns-use-cases/sagas/sagas-typescript/README.md
@@ -1,18 +1,15 @@
-# Restate Sample Patterns
+# SAGAs / Compensations
 
-A collections of useful patterns in distributed applications, and how to
-implement them with [Restate](https://github.com/restatedev/).
-This is a continuously evolving list.
+An example of a trip reservation workflow, using the SAGAs pattern to
+undo previous steps in csase of an error.
+This is a minimal version of the holiday reservation demo in the
+[Restate Holiday Repository](https://github.com/restatedev/restate-holiday).
 
-All patterns are described in one file and have a comment block at the top
-that explains them.
+Durable Execution's guarantee to run code to the end in the presence
+of failures, and to deterministically recover previous steps from the
+journal, makes SAGAs easy.
+Every step pushes a compensation action (an undo operation) to a stack.
+in the case of an error, those operations are run.
 
-### List of Patterns
-
-- [**Durable Event-based Asynchronous Communication as RPC**](src/async_calls_as_rpc.ts)
-- [**Idempotency Tokens**](src/deterministic_idempotency_tokens.ts)
-- [**Single-writer Concurrency**](src/single_writer_concurrency.ts)
-- [**Dual Writes**](src/dual_writes.ts)
-- [**Distributed Locks**](src/distributed_locks.ts)
-- [**Idempotent DynamoDB Updates**](src/dynamo_db_idempotency.ts)
-- [**Compensations**](src/compensations.ts)
+The main requirement is that steps are implemented as journalled
+operations, like `ctx.sideEffect()` or `ctx.rpc()`.
